% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{head1}
\alias{head1}
\title{Print 1:6 Rows and Cols}
\usage{
head1(obj, i = 6, j = 6)
}
\arguments{
\item{obj}{data.frame or matrix}

\item{i}{rows}

\item{j}{cols}
}
\value{
data.frame or matrix
}
\description{
Print 1:6 Rows and Cols
}
\examples{
head1(mtcars)
}
