% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head1.R
\name{head1}
\alias{head1}
\alias{head1.default}
\alias{head1.matrix}
\title{Print 1:6 Rows and Cols}
\usage{
head1(obj, ...)

\method{head1}{default}(obj, ...)

\method{head1}{matrix}(obj, n = 6, m = 6, ...)
}
\arguments{
\item{obj}{data.frame or matrix}

\item{...}{args passed to \code{\link[utils:head]{utils::head()}}}

\item{n}{rows}

\item{m}{cols}
}
\value{
data.frame or matrix
}
\description{
If a matrix/df is too big, Rstudio will try to print out everything and
freeze. This is a wrapper around \code{head(x, c(n, m))}.
}
\examples{
head1(mtcars)
}
